refactor: move createProtectedRouter to auth-context.ts
