+++ cli/template/addons/trpc/protected-example-router.ts
@@ -1,4 +1,4 @@
-import { createProtectedRouter } from "./protected-router";
+import { createProtectedRouter } from "./context";
 
 // Example router with queries that can only be hit if the user requesting is signed in
 export const protectedExampleRouter = createProtectedRouter()
